backend/requirements.txt
fastapi==0.116.1
uvicorn==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
shortuuid==1.0.13
slowapi==0.1.9
redis==6.2.0
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.28.1
psutil==7.0.0
pydantic[email]==2.11.7 